# HotelManagementAPI/Datas/Models/Room.cs
using System;
using System.Collections.Generic;

namespace HotelManagementAPI.Datas.Models
{
    public partial class Room
    {
        public Room()
        {
            Bookings = new HashSet<Booking>();
            Reviews = new HashSet<Review>();
        }

        public string Id { get; set; } = null!;
        public string? Name { get; set; }
        public string? Type { get; set; }
        public decimal? Price { get; set; }
        public decimal? Area { get; set; }
        public string? Description { get; set; }
        public string? Status { get; set; }

        public virtual ICollection<Booking> Bookings { get; set; }
        public virtual ICollection<Review> Reviews { get; set; }
    }
}